Move the logic for yaml and json to their base
