+++ insights/formats/__init__.py
@@ -76,10 +76,18 @@ class EvaluatorFormatterAdapter(FormatterAdapter):
     """
     Impl = None
 
+    @staticmethod
+    def configure(p):
+        """ Override to add arguments to the ArgumentParser. """
+        p.add_argument("-F", "--fail-only", help="Show FAIL results only. Conflict with '-m' or '-f', will be dropped when using them together", action="store_true")
+
     def __init__(self, args=None):
         if args:
             hn = "insights.combiners.hostname"
             args.plugins = ",".join([args.plugins, hn]) if args.plugins else hn
+            if args.fail_only:
+                print('Options conflict: -f and -F, drops -F', file=sys.stderr)
+                args.fail_only = False
 
     def preprocess(self, broker):
         self.formatter = self.Impl(broker)
