+++ insights/formats/_json.py
@@ -1,6 +1,4 @@
-from __future__ import print_function
 import json
-import sys
 
 from insights.core.evaluators import SingleEvaluator
 from insights.formats import EvaluatorFormatterAdapter
@@ -13,12 +11,3 @@ class JsonFormat(SingleEvaluator):
 
 class JsonFormatterAdapter(EvaluatorFormatterAdapter):
     Impl = JsonFormat
-
-    @staticmethod
-    def configure(p):
-        p.add_argument("-F", "--fail-only", help="Show FAIL results only. Conflict with '-m' or '-f', will be dropped when using them together", action="store_true")
-
-    def __init__(self, args):
-        super(JsonFormatterAdapter, self).__init__(args)
-        if args.fail_only:
-            print('Options conflict: -f and -F, drops -F', file=sys.stderr)
